Improvements
-Changed daynighttimer to change the suns rotation multiple times a frame, while keeping the movement of [time] steady, to simulate smooth movement.
-Edited the enemy class to drop post action meshes of appropriate scale. Scale was not included in the initial manifestation.
-Created the resourcespawner class and created base functionality less the inclusion of a defualt value spreadsheet to set values needed by the inventory class

+++ DayNightCycle.h
@@ -36,7 +36,7 @@ public:
 	//BP_Sky_Sphere_C* SKySphere;
 
 	int hour;
-	int min;
+	float min;
 	int sec;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = sun)
